perf(ui): defer the roster cache reader off the boot path
Load roster validation and IndexedDB reads on demand while preserving cache-generation fences and warm-roster settlement. Separate the Skills status RPC from its lazy management operations to keep startup JavaScript within the unchanged budget.

Consolidate synchronous warm-state clearing and resolve the rebased warm-boot lint failures without changing behavior or check limits.
