+++ ui/src/app/bootstrap-warm-boot.ts
@@ -3,6 +3,17 @@ import { markPrewarmedChatSnapshotReady } from "../pages/chat/session-snapshot-p
 import { clearBootRecords } from "./boot-record.ts";
 import type { ApplicationGateway } from "./gateway.ts";
 
+export function clearWarmBootState(): void {
+  // The boot record gates the next warm boot, so it must be gone before any
+  // await: a reload during the lazy IndexedDB cleanup must fail closed.
+  clearBootRecords();
+  // Invalidate visible history and its cursor before pane subscribers resume startup.
+  void clearStoredChatSnapshots();
+  void import("../lib/sessions/session-roster-cache.runtime.ts").then(({ clearCachedBootState }) =>
+    clearCachedBootState(),
+  );
+}
+
 export function subscribeWarmBootConnection(
   gateway: ApplicationGateway,
   profileId: string | null | undefined,
@@ -22,14 +33,7 @@ export function subscribeWarmBootConnection(
     const profileMismatch = pendingBootProfileId !== (snapshot.selfUser?.id ?? null);
     pendingBootProfileId = undefined;
     if (profileMismatch) {
-      // The boot record gates the next warm boot, so it must be gone before any
-      // await: a reload during the lazy IndexedDB cleanup must fail closed.
-      clearBootRecords();
-      // Invalidate visible history and its cursor before pane subscribers resume startup.
-      void clearStoredChatSnapshots();
-      void import("../lib/sessions/session-roster-cache.runtime.ts").then(
-        ({ clearCachedBootState }) => clearCachedBootState(),
-      );
+      clearWarmBootState();
     }
   });
 }
