+++ ui/src/app/bootstrap.warm-boot.test.ts
@@ -28,7 +28,7 @@ describe("warm boot profile validation", () => {
       { cachedProfileId: "profile-a", profileId: "profile-a", clears: 0 },
       { cachedProfileId: null, profileId: null, clears: 0 },
       { cachedProfileId: "profile-a", profileId: "profile-b", clears: 0, credentialsChanged: true },
-    ].map((entry) => ({ ...entry, pathname: "/chat", warmBoot: true })),
+    ].map((entry) => Object.assign(entry, { pathname: "/chat", warmBoot: true })),
     ...["/focus/terminal", "/approve/exec%3A1"].map((pathname) => ({
       cachedProfileId: "profile-a",
       profileId: "profile-b",
