+++ ui/src/app/gateway-store.ts
@@ -29,8 +29,7 @@ import { setAvatarGatewayOrigin } from "../lib/identity-avatar-context.ts";
 import { resolveSessionKey } from "../lib/sessions/index.ts";
 import { readSessionDefaults } from "../lib/sessions/session-key.ts";
 import { generateUUID } from "../lib/uuid.ts";
-import { clearStoredChatSnapshots } from "../pages/chat/session-snapshot-invalidation.runtime.ts";
-import { clearBootRecords } from "./boot-record.ts";
+import { clearWarmBootState } from "./bootstrap-warm-boot.ts";
 import type {
   ApplicationGateway,
   ApplicationGatewayConnectOptions,
@@ -377,11 +376,7 @@ export function createApplicationGateway(
     const retiredEventLog = credentialsChanged ? eventLog.resetConnection() : null;
     if (credentialsChanged) {
       connectionRevision += 1;
-      clearBootRecords();
-      void clearStoredChatSnapshots();
-      void import("../lib/sessions/session-roster-cache.runtime.ts").then(
-        ({ clearCachedBootState }) => clearCachedBootState(),
-      );
+      clearWarmBootState();
     }
     // Only a gateway URL that differs from the current connection counts as an
     // explicit selection. The login gate always resubmits its prefilled URL, so
